pyproject.toml
[project]
name = "trade-py"
version = "0.1.0"
dependencies = [
  "absl-py",
  "numpy",
  "pandas >=2.0",
]
